- Grades working + fixed insert into DB - Fixed layouts - Fixed Profile view - Deleted unused files - Params for grades and catalog still not fixed - Added stored procedure for computing final grade

+++ QuizWebApplication/SQL/Tables/Grades.sql
@@ -3,9 +3,8 @@ CREATE TABLE Grades
 		Id int primary key identity(1,1),
 		UserId INT,
 		CategoryId INT,
-		Grade INT
+		Grade FLOAT
     ) 
-
 
 ALTER TABLE   
 	Grades
